:sparkles: main, player state 통합 및 state로 분리
feature/solo의 구조를 기반으로 main_state와 player_state를 합쳤습니다. 또한 공의 방향 계산과정을 state1에 전부 넣었습니다.
