+++ src/constants.rs
@@ -20,5 +20,6 @@ pub enum StateTransition {
     None,
     ToTitle,
     ToMain,
-    ToPlayer,
+    Multi,
+    Solo,
 }
